refactor: move checkQueryReturn() to CheckQueryReturnTrait
